requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Werkzeug>=3.0.1
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
web3>=6.11.0
py-solc-x>=2.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
